requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.50.0
django-autoslug==1.9.8
django-cors-headers==3.12.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
idna==3.3
Markdown==3.3.7
oauthlib==3.2.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
